# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server Petstore server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/common/media/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get medias",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common-admin"
                ],
                "summary": "get medias",
                "responses": {}
            }
        },
...
        "/admin/common/partner/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list partners",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common-admin"
                ],
                "summary": "list partners",
                "responses": {}
            }
        },
...
        "/admin/company/capasity/create": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "create capasity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-admin"
                ],
                "summary": "create capasity",
                "parameters": [
                    {
                        "description": "payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types_about_company.CapasityPayload"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/admin/company/capasity/list": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "list capasity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-admin"
                ],
                "summary": "list capasity",
                "responses": {}
            }
        },
        "/admin/company/capasity/{capasityId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "get capasity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-admin"
                ],
                "summary": "get capasity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "capasity id",
                        "name": "capasityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/company/capasity/{capasityId}/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "delete capasity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-admin"
                ],
                "summary": "delete capasity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "capasity id",
                        "name": "capasityId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/admin/company/capasity/{capasityId}/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "update capasity",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company-admin"
                ],
                "summary": "update capasity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "capasity id",
                        "name": "capasityId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update capasity payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types_about_company.CapasityPayload"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/contact_us/create": {
            "post": {
                "description": "create contact us",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common"
                ],
                "summary": "create contact us",
                "parameters": [
                    {
                        "description": "Contact Us Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types_common.ContactCreatePayload"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/contact_us_footer/create": {
            "post": {
                "description": "create contact footer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "common"
                ],
                "summary": "create contact footer",
                "parameters": [
                    {
                        "description": "contact us footer create payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/types_common.ContactUsFooterPayload"
                        }
                    }
                ],
                "responses": {}
            }
        },
...
    },
    "definitions": {
        "types_about_company.CapasityPayload": {
            "type": "object",
            "properties": {
                "nameEn": {
                    "type": "string"
                },
                "nameRu": {
                    "type": "string"
                },
                "nameUz": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "types_common.ContactCreatePayload": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "comment": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                }
            }
        },
        "types_common.ContactUsFooterPayload": {
            "type": "object",
            "required": [
                "email",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}